+++ src/test/kotlin/org/nemwiz/jiracommitmessage/services/JiraCommitMessagePluginTest.kt
@@ -40,7 +40,7 @@ class JiraCommitMessagePluginTest : BasePlatformTestCase() {
         PluginSettingsState.instance.state.jiraProjectKeys = listOf(dummyProjectKey1, dummyProjectKey2)
         PluginSettingsState.instance.state.isAutoDetectJiraProjectKey = false
 
-        val mockBranch1 = "feat/${dummyProjectKey2}-8172"
+        val mockBranch1 = "hello/${dummyProjectKey2}-8172"
 
         val plugin = project.service<JiraCommitMessagePlugin>()
 
@@ -48,7 +48,7 @@ class JiraCommitMessagePluginTest : BasePlatformTestCase() {
 
         assertEquals("(${dummyProjectKey2}-8172)", commitMessage1)
 
-        val mockBranch2 = "feat/${dummyProjectKey1}_123-foo"
+        val mockBranch2 = "hello/${dummyProjectKey1}_123-foo"
 
         val commitMessage2 = plugin.getCommitMessageFromBranchName(mockBranch2)
 
@@ -200,5 +200,108 @@ class JiraCommitMessagePluginTest : BasePlatformTestCase() {
         val commitMessage = plugin.getCommitMessageFromBranchName(mockBranch)
 
         assertEquals("", commitMessage)
+    }
+
+    fun testReturnConventionalCommitInfixWhenConvetionalCommitsOptionIsOn() {
+        val dummyProjectKey1 = "LEGOS"
+
+        PluginSettingsState.instance.state.jiraProjectKeys = listOf(dummyProjectKey1)
+        PluginSettingsState.instance.state.isAutoDetectJiraProjectKey = false
+        PluginSettingsState.instance.state.isConventionalCommit = true
+        PluginSettingsState.instance.pluginState.messageInfixType = InfixType.NO_INFIX.type
+        PluginSettingsState.instance.state.messageWrapperType = MessageWrapperType.ROUND.type
+
+        val mockBranch1 = "feat/awesome-LEGOS-123"
+
+        val plugin = project.service<JiraCommitMessagePlugin>()
+
+        val commitMessage1 = plugin.getCommitMessageFromBranchName(mockBranch1)
+
+        assertEquals("feat(${dummyProjectKey1}-123)", commitMessage1)
+
+        val mockBranch2 = "chore-awesome-LEGOS-123"
+
+        val commitMessage2 = plugin.getCommitMessageFromBranchName(mockBranch2)
+
+        assertEquals("chore(${dummyProjectKey1}-123)", commitMessage2)
+    }
+
+    fun testReturnConventionalCommitInfixWhenConvetionalCommitsOptionIsOnAndAutomaticDetectIsOn() {
+        PluginSettingsState.instance.state.isAutoDetectJiraProjectKey = true
+        PluginSettingsState.instance.state.isConventionalCommit = true
+        PluginSettingsState.instance.pluginState.messageInfixType = InfixType.COLON.type
+        PluginSettingsState.instance.state.messageWrapperType = MessageWrapperType.ROUND.type
+
+        val mockBranch1 = "fix/awesome-PLUGIN-831"
+
+        val plugin = project.service<JiraCommitMessagePlugin>()
+
+        val commitMessage1 = plugin.getCommitMessageFromBranchName(mockBranch1)
+
+        assertEquals("fix(PLUGIN-831):", commitMessage1)
+
+        val mockBranch2 = "docs-awesome-PLUGIN-831"
+
+        val commitMessage2 = plugin.getCommitMessageFromBranchName(mockBranch2)
+
+        assertEquals("docs(PLUGIN-831):", commitMessage2)
+    }
+
+    fun testDoesNotAddConventionalCommitInfixWhenConventionalCommitsAreNotOn() {
+        PluginSettingsState.instance.state.isAutoDetectJiraProjectKey = true
+        PluginSettingsState.instance.state.isConventionalCommit = false
+        PluginSettingsState.instance.pluginState.messageInfixType = InfixType.NO_INFIX.type
+        PluginSettingsState.instance.state.messageWrapperType = MessageWrapperType.BOX.type
+
+        val mockBranch = "test/awesome-PLUGIN-831"
+
+        val plugin = project.service<JiraCommitMessagePlugin>()
+
+        val commitMessage = plugin.getCommitMessageFromBranchName(mockBranch)
+
+        assertEquals("[PLUGIN-831]", commitMessage)
+    }
+
+    fun testOnlyAddsInfixFromTheSupportedConventionalCommitInfix() {
+        PluginSettingsState.instance.state.isAutoDetectJiraProjectKey = true
+        PluginSettingsState.instance.state.isConventionalCommit = true
+        PluginSettingsState.instance.pluginState.messageInfixType = InfixType.NO_INFIX.type
+        PluginSettingsState.instance.state.messageWrapperType = MessageWrapperType.ROUND.type
+
+        val branches = listOf(
+            "test/PROJ-123",
+            "ci/PROJ-123",
+            "chore/PROJ-123",
+            "build/PROJ-123",
+            "fix/PROJ-123",
+            "feat/PROJ-123",
+            "docs/PROJ-123",
+            "perf/PROJ-123",
+            "refactor/PROJ-123",
+            "style/PROJ-123",
+            "unknown/PROJ-123",
+            "dummy/PROJ-123",
+        )
+
+
+        val plugin = project.service<JiraCommitMessagePlugin>()
+
+        val commitMessages = branches.map { b ->  plugin.getCommitMessageFromBranchName(b)}
+
+        assertEquals(branches.size, commitMessages.size)
+
+        assertFalse(commitMessages.contains("dummy(PROJ-123)"))
+        assertFalse(commitMessages.contains("unknown(PROJ-123)"))
+
+        assertTrue(commitMessages.contains("test(PROJ-123)"))
+        assertTrue(commitMessages.contains("ci(PROJ-123)"))
+        assertTrue(commitMessages.contains("chore(PROJ-123)"))
+        assertTrue(commitMessages.contains("build(PROJ-123)"))
+        assertTrue(commitMessages.contains("fix(PROJ-123)"))
+        assertTrue(commitMessages.contains("feat(PROJ-123)"))
+        assertTrue(commitMessages.contains("docs(PROJ-123)"))
+        assertTrue(commitMessages.contains("perf(PROJ-123)"))
+        assertTrue(commitMessages.contains("refactor(PROJ-123)"))
+        assertTrue(commitMessages.contains("style(PROJ-123)"))
     }
 }