+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 ## [Unreleased]
 
+### Added
+- Support for conventional commits based on branch name
+
 ## [3.2.0] - 2023-01-27
 
 ### Added